pyproject.toml
[tool.poetry]
name = "transactions-analyzer"
version = "0.1.0"
description = ""
authors = ["Mwathi <mwathibrian7@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
amqp = "5.2.0"
async-timeout = "4.0.3"
billiard = "4.2.0"
blinker = "1.7.0"
camelot-py = "0.9.0"
celery = "5.3.6"
cffi = "1.16.0"
chardet = "5.2.0"
charset-normalizer = "3.3.2"
click = "8.1.7"
click-didyoumean = "0.3.1"
click-plugins = "1.1.1"
click-repl = "0.3.0"
cryptography = "41.0.7"
docopt = "0.6.2"
et-xmlfile = "1.1.0"
flask = "3.0.0"
flask-cors = "4.0.0"
importlib-metadata = "7.0.1"
itsdangerous = "2.1.2"
jinja2 = "3.1.3"
kombu = "5.3.7"
markupsafe = "2.1.3"
numpy = "1.26.3"
opencv-python = "4.8.0.76"
openpyxl = "3.1.2"
pandas = "2.0.0"
pdfminer-six = "20220524"
prompt-toolkit = "3.0.43"
py-pdf-parser = "0.10.2"
pycparser = "2.21"
pypdf2 = "2.12.0"
python-dateutil = "2.8.2"
pytz = "2023.3.post1"
redis = "5.0.4"
six = "1.16.0"
typing-extensions = "4.9.0"
tzdata = "2023.4"
vine = "5.1.0"
wand = "0.6.10"
wcwidth = "0.2.13"
werkzeug = "3.0.1"
zipp = "3.17.0"
load-dotenv = "^0.1.0"
gunicorn = "^23.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
